+++ steel-core/src/entity/mob_effect/mod.rs
@@ -69,7 +69,7 @@ pub trait InstantaneousMobEffect: MobEffectBehavior {
         amplifier: i32,
         direct_entity: Option<i32>,
         causing_entity: Option<i32>,
-        scale: f32,
+        scale: f64,
     ) {
         let _ = (direct_entity, causing_entity, scale);
         self.apply_effect_tick(world, user, amplifier);
